feat: metodos de adicao e subtracao de saldo

+++ src/main/java/com/anadev/account_service/dto/AccountRequest.java
@@ -4,6 +4,7 @@ import com.anadev.account_service.entity.enums.TypeAccount;
 import com.anadev.account_service.entity.enums.TypeCurrency;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 @Builder
@@ -11,6 +12,6 @@ public record AccountRequest(
         String name,
         TypeAccount typeAccount,
         TypeCurrency currency,
-        Double currentBalance,
-        Double monthlyLimit
+        BigDecimal currentBalance,
+        BigDecimal monthlyLimit
 ){}
