+++ src/main/java/com/anadev/account_service/dto/AccountResponse.java
@@ -4,6 +4,7 @@ import com.anadev.account_service.entity.Account;
 import com.anadev.account_service.entity.enums.TypeAccount;
 import com.anadev.account_service.entity.enums.TypeCurrency;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 
@@ -12,8 +13,8 @@ public record AccountResponse(
         String name,
         TypeAccount typeAccount,
         TypeCurrency currency,
-        Double currentBalance,
-        Double monthlyLimit)
+        BigDecimal currentBalance,
+        BigDecimal monthlyLimit)
 {
     public static AccountResponse fromEntity(Account account){
         return new AccountResponse(
